Remove public_key argument from glome_tag function.
The argument was redundant with the private_key argument, and introduced a potential for usage erros where the public key would (accidentally) not match the private key.

This entails that the public key needs to be recalculated with every tag. Fortunately, this operation is relatively cheap, and the overall increase in glome_tag runtime does not seem to exceed 2.5%.

PiperOrigin-RevId: 322184602