libglome: Unit tests for documented test vectors
This change adds unit tests using the GLib testing framework.
The two tests that are added are the two that are present in the
protocol documentation to ensure compatibility.

This change also introduces a meson build switch to enable building
test suites. This switch is not yet used as an internal patch relies on
some specific layout of the meson.build right now.

Signed-off-by: Christian Svensson <bluecmd@google.com>