meson.build: force finding an absolute path for cli/test.sh
Currently meson ends up resolving the script path to the literal string
"cli/test.sh", which breaks when trying to run the tests from a
different directory than the source directory. A common reason one might
want to do this is if running the tests via "ninja test" instead of
"meson test".

TESTED=Tests successfully run with "ninja test" with this change.