KeyFetcher signature to use ptr
This improves the impedance between the glome.PrivateKey object and
helper functions and login.Server's KeyFetcher. This makes the error
case and using the helper functions smoother.

In the case of a key not being found the return of the key fetcher is
changed to be a success with a nil value returned. This means that only
internal errors should lead to error conditions for the key fetcher.

Signed-off-by: Christian Svensson <bluecmd@google.com>