Support compilation with both old and new pam_wrapper
pam_wrapper changed the signature of run_pamtest to allow for five
arguments - to enable reusing an existing pam_handle. However this
breaks the API in a hard to check for way. The pkg-config file also does
not specify a version. So this tests if compilation with the old
macro signature succeeds and if so conditionally uses the old syntax in
the code.