Rename `url-prefix` to `prompt`
Since not all users are going to run web server to generate authcodes,
it might be misleading to use the term 'URL'.

The leading slash can also be dropped from the challenge format since it
does not play a significant role.

The code still uses the term 'URL' but this will be addressed in a
subsequent change for better readability.

Issue #125