State that devices require randomness for the protocol to work
The ephemeral key generated by Alice must not be predictable. The protocol should not leak anything about the key right now as it gets munged into the elliptic curve calculation, but if the PRNG seed is known or it is seeded from predictable information (like time), then it might be possible to do an offline attack on the key space. The timeout helps us here, but only if the key space is large enough.