Implement GLOME login INI config parsing in Go.
This implements INI parsing internally, rather than pulling in a library because
we want to exactly match the behaviour of the C library, and our INI-parsing has
a few quirks (as do all INI parsers...) - for instance, we only support comments
if they're on their own line.

Additionally, the code that does the INI parsing is relatively small; the bulk of
the code is actually parsing things into application specific structs, which we
would need to do anyway. Possibly this is ripe for switching to TOML, but we'd still
have to e.g. parse string encoded keys to binary/validate that the options set
are compatible with each other.

Fixes #168.