Avoid calling openlog when using GLOME as a library.
This causes log lines to always use `glome-login` as the ident,
which isn't necessarily desirable.

We can just not call openlog: if we don't call it, syslog will
initialise with some default values (i.e. not including the PID).
To avoid this then logging to the wrong facility we update our
own calls to syslog to explicitly set the facility.

PAM doesn't need a facility; it appears to just use LOG_AUTHPRIV
internally.