Disable bracketed paste
In some settings we seem to end up with users' terminals being confused
and willing to send bracketed paste escape codes, which then cause
authentication to fail. Even worse it looks like the special characters
are not displayed in some shells, thus confusing users as to what just
happened.

Although to be honest: I have not found the specific environment where
this breaks in a reproducible fashion and this is effectively defence
in depth.